car-parking-python/requirements.txt
pynput
pyserial
requests